move chat stuff over
